+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Buttons to generate release-notes (before or after PR merge)
 - Send Git Provider & Ticketing provider auth info when running sf hardis commands (ex: Generate release notes)
 - Optimize user login and reconnect experiences
+- Add backpromote command (beta)
 
 ## [7.7.1] 2026-04-30